[wb-1901-tabs-with-tooltips-popovers] move examples with tooltips and popovers back to main component stories instead of playtesting stories
